testing/web-platform/tests/workers/WorkerNavigator_appVersion.htm

<!DOCTYPE html>
<html>
<head>
<title> WorkerNavigator appVersion </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id=log></div>
<script>
    
    var FileName = "./support/WorkerNavigator.js";
    var description = "WorkerNavigator appVersion: Returns the version of the browser.";
    
    var t = async_test("Test Description: " + description);
                  
    var worker = new Worker(FileName);

    worker.onmessage = t.step_func(function(evt)
    {
        assert_equals(evt.data.appVersion, window.navigator.appVersion);
        t.done();
    });
</script>
</body>
</html>
